server/users_app/urls.py
from django.urls import path
from .views import RegisterView, EmailCookieTokenObtainPairView
from rest_framework_simplejwt.views import TokenRefreshView

urlpatterns = [
    path("register/", RegisterView.as_view(), name="user-register"),
    # JWT endpoints
    path("token/", EmailCookieTokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
]
